Update Method.pm
So there were a few problems here.  First we were double UTF-8 encoding our length (its already UTF-8 encoded so no need to do it again), and second the call to length gives you the length in characters not in bytes.  To get the bytes use the bytes::length method